站长信息功能晚上

+++ Common/Common/function.php
@@ -5,7 +5,7 @@ function pwdHash($password, $type = 'md5') {
 }
 
 //递归删除某目录下所有文件
-function delFileUnderDir($dirName) {
+function delFileUnderDir($dirName, $keep_file = null) {
     $flag = false;
     if ($handle = opendir("$dirName")) {
         while (false !== ( $item = readdir($handle) )) {
@@ -13,7 +13,13 @@ function delFileUnderDir($dirName) {
                 if (is_dir("$dirName/$item")) {
                     delFileUnderDir("$dirName/$item");
                 } else {
-                    unlink("$dirName/$item");
+                    if ($keep_file !== null) {
+                        if ($keep_file != "$dirName/$item") {
+                            unlink("$dirName/$item");
+                        }
+                    } else {
+                        unlink("$dirName/$item");
+                    }
                 }
             }
         }
